Create clean.py
Checks each headline in outHeadlines.txt to see if it matches to one of the tickers from outData.txt. Headlines do not have the tickers in them and instead have full names of companies, so next I need to find a way to relate the names to the tickers before checking the headline.